requirements.txt
django
flake8
djangorestframework
drf_yasg
django-cors-headers